Consolidate `ShortcakeApp` and `ShortcakeClient`. Move routing code
